+++ installation-bundle/src/ContaoInstallationBundle.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of Contao.
+ *
+ * Copyright (c) 2005-2015 Leo Feyer
+ *
+ * @license LGPL-3.0+
+ */
+
+namespace Contao\InstallationBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+/**
+ * Configures the Contao installation bundle.
+ *
+ * @author Leo Feyer <https://github.com/leofeyer>
+ */
+class ContaoInstallationBundle extends Bundle
+{
+}